requirements.txt
django==2.2
djangorestframework
django-model-utils
